add win test
